Enter city and country, and then displays a message stating where they live.
This C++ program prompts the user to enter their city and country, then displays a message with that information. It demonstrates basic input and output operations using `cin` and `cout`. And also in this code we didn't use the line of 'using namespace std; line .So we should have to add std:: fpr all the variables.